requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
Django==3.2.3
djangorestframework==3.12.4
kombu==5.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
